Consolidate scheduling module and improve captcha flow
